Fix bug regarding the streaming response.
`websocket_generate_stream` was trying to manipulate
the response using dict, but we changed the response to
be a pydantic model. Then we need to either transform
to a dict before, or use the dot notation.

Transforming to `.dict()` before is better as we send it to
the websocket as a dict.